create edit action buttons component
